Add stores::Policy and start stubbing in auth{z,n}
This adds GitHub oauth login, a "Policy" system for constructing aggregate
implementations of various required registry services, and the beginning of
auth.

I've split authentication between "TokenAuthorizer" and "Authenticator"; I
think there still needs to be a "UserStorage" trait as well. The idea is that
Authenticator implements the CLI web auth using an OAuth flow, maintaining an
in-memory map of "login registration sessions". Once the oauth flow is
complete, we register the resulting internal User object with the appropriate
login registration session. The next time the CLI polls for a user, the
endpoint finds the user associated with the login registration session. The
handler then registers the user with the token authorizer.

And this is where UserStorage would come in: between those two steps, we should
store the user _somewhere_ (and validate that we actually want to accept this
user.)

Additionally: add a Configurator type to hold general configuration data. This
is a little bit of a weird one! To explain: I've made some of the
more-sensitive values async in order to provide a Vault or AWS Secrets Manager
Configurator in the future. Yanno: for stuff you don't necessarily want
floating around memory indefinitely.

I've added (and removed) the extism plugin. I think it will make its triumphant
return soon -- I want to run accept/reject/transform plugins on package version
modification (whether or not that's via mirroring) as well as user
registration.

Anyway, my eyes are crossing as I write this.